fix: Extract Trello colors into a separate file
