+++ js/src/admin/components/ExtensionSettingsPage.tsx
@@ -9,6 +9,7 @@ import tagLabel from 'flarum/tags/helpers/tagLabel';
 import sortTags from 'flarum/tags/utils/sortTags';
 import saveSettings from 'flarum/admin/utils/saveSettings';
 import type Mithril from 'mithril';
+import colors from '../../common/colors';
 
 export interface Board {
   organization(): string;
@@ -213,20 +214,6 @@ export default class TrelloSettingsPage extends ExtensionPage {
   content() {
     function displayLabel(label: {}) {
       let attrs = {};
-
-      // Trello color choices
-      const colors = {
-        green: '#61bd4f',
-        yellow: '#f2d600',
-        orange: '#ff9f1a',
-        red: '#eb5a46',
-        purple: '#c377e0',
-        blue: '#0079bf',
-        sky: '#00c2e0',
-        lime: '#51e898',
-        pink: '#ff78cb',
-        black: '#344563',
-      };
 
       attrs.style = {};
       attrs.className = 'TagLabel ';
