+++ js/src/common/colors.ts
@@ -1,0 +1,15 @@
+// Trello color choices
+const colors = {
+  green: '#61bd4f',
+  yellow: '#f2d600',
+  orange: '#ff9f1a',
+  red: '#eb5a46',
+  purple: '#c377e0',
+  blue: '#0079bf',
+  sky: '#00c2e0',
+  lime: '#51e898',
+  pink: '#ff78cb',
+  black: '#344563',
+};
+
+export default colors;
